feat: building emission form_data
